requirements.txt
python-telegram-bot>=20.7,<21.0
google-generativeai>=0.3.0,<0.6.0
pdfplumber>=0.10.0,<0.12.0
reportlab>=4.0.0,<4.2.0
python-dotenv>=1.0.0,<1.2.0
langdetect>=1.0.9,<1.1.0
